✅ FEAT: listen for project promise to finish loading
