+++ src/pages/Home.js
@@ -2,11 +2,12 @@ import Header from '../components/header/Header';
 import AboutMe from '../components/about/About';
 import Projects from '../components/projects/Projects';
 
-const Home = () => (
+// eslint-disable-next-line react/prop-types
+const Home = ({ projects }) => (
   <div>
     <Header />
     <AboutMe />
-    <Projects />
+    <Projects projects={projects} />
   </div>
 );
